Correção de bugs do relatório para adequar à história

+++ src/SME.SR.Infra/Dtos/Relatorios/SondagemPortugues/RelatorioSondagemPortuguesConsolidadoLeituraPlanilhaRespostaDto.cs
@@ -9,7 +9,7 @@ namespace SME.SR.Infra
     {
         public string Resposta { get; set; }
         public int Quantidade { get; set; }
-        public double Percentual { get; set; }
+        public decimal Percentual { get; set; }
         public int Total { get; set; }
     }
 }
